Log form data or message to the console on save
If a product type is selected, the product data will be logged to the console, otherwise a message will be logged indicating that the user must select a product type. The user is not prevented from sending a blank form, which should be addressed. The form is reset upon submission. A bug was fixed that logged messages / data twice, because TheButton component emitted a  event, which conflicted with the browser's native  event. After renaming it , the message / data is now logged only once.