Basic form validation
Form validation is currently not perfect. 1. The message ‘Please select a product type’ is displayed while warning styles are applied to the blank input fields (warnings mismatch). Find a way to highlight the select dropdown to draw attention to the element that is the object of the message, or display all messages at the same time.
2. Prevent form from being sent without filling in inputs of type number. Currently, only blank inputs of type text prevent the form submission.